adicionadas validacoes

+++ PontoApp/Controllers/EmployeesController.cs
@@ -1,5 +1,4 @@
-#nullable disable
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -23,13 +22,15 @@ namespace PontoApp.Controllers
         // GET: Employees
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Employee.ToListAsync());
+              return _context.Employee != null ? 
+                          View(await _context.Employee.ToListAsync()) :
+                          Problem("Entity set 'Context.Employee'  is null.");
         }
 
         // GET: Employees/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Employee == null)
             {
                 return NotFound();
             }
@@ -55,7 +56,7 @@ namespace PontoApp.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Name,Email,Password,Cpf,Rg,Empresa,Telefone")] Employee employee)
+        public async Task<IActionResult> Create([Bind("Id,Name,Email,Password,Cpf,Rg,Empresa,Telefone,Cidade")] Employee employee)
         {
             if (ModelState.IsValid)
             {
@@ -69,7 +70,7 @@ namespace PontoApp.Controllers
         // GET: Employees/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Employee == null)
             {
                 return NotFound();
             }
@@ -87,7 +88,7 @@ namespace PontoApp.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Email,Password,Cpf,Rg,Empresa,Telefone")] Employee employee)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Email,Password,Cpf,Rg,Empresa,Telefone,Cidade")] Employee employee)
         {
             if (id != employee.Id)
             {
@@ -120,7 +121,7 @@ namespace PontoApp.Controllers
         // GET: Employees/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Employee == null)
             {
                 return NotFound();
             }
@@ -140,15 +141,23 @@ namespace PontoApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Employee == null)
+            {
+                return Problem("Entity set 'Context.Employee'  is null.");
+            }
             var employee = await _context.Employee.FindAsync(id);
-            _context.Employee.Remove(employee);
+            if (employee != null)
+            {
+                _context.Employee.Remove(employee);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool EmployeeExists(int id)
         {
-            return _context.Employee.Any(e => e.Id == id);
+          return (_context.Employee?.Any(e => e.Id == id)).GetValueOrDefault();
         }
     }
 }
